pyproject.toml
[tool.poetry]
name = "chat-service"
version = "0.1.0"
description = ""
authors = ["Vachagan <vachagan.grigoryan@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.217.0"}
uvicorn = "^0.25.0"
websockets = "^12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
